added check for existing system config when a feature is enabled

+++ abstracto-application/abstracto-modules/moderation/moderation-impl/src/main/java/dev/sheldan/abstracto/moderation/listener/WarnDecayConfigListener.java
@@ -18,6 +18,6 @@ public class WarnDecayConfigListener implements ServerConfigListener {
 
     @Override
     public void updateServerConfig(AServer server) {
-        configManagementService.createIfNotExists(server.getId(), "decayDays", decayDays.longValue());
+        configManagementService.createIfNotExists(server.getId(), "decayDays", decayDays);
     }
 }
